streamlit_app/requirements.txt
absl-py==1.4.0
accelerate==0.31.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.1
algorithms==0.1.4
altair==4.2.2
anaconda-navigator==2.3.2
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
audioread==3.0.1
backports.weakref==1.0.post1
bkcharts==0.2
blinker==1.5
blis==0.7.10
Bottleneck==1.4.0
branca==0.6.0
brotlipy==0.7.0
cachetools==5.3.0
catalogue==2.0.9
certifi==2021.5.30
chardet==3.0.4
cleantext==1.1.4
clyent==1.2.2
components==1.2.8
comtypes==1.1.10
conda==22.9.0
conda-build==3.22.0
conda-repo-cli==1.0.20
conda-verify==3.4.2
confection==0.1.0
constantly==15.1.0
cssselect==1.1.0
customtkinter==5.2.1
cymem==2.0.7
cytoolz==0.11.0
daal4py==2021.6.0
darkdetect==0.8.0
datasets==2.14.1
datashape==0.5.4
discord.py==2.3.2
distlib==0.3.6
duckdb==0.8.1
emoji==2.8.0
et-xmlfile==1.1.0
evaluate==0.4.2
exceptiongroup==1.1.1
extra-streamlit-components==0.1.56
flasgger==0.9.5
Flask==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
flatbuffers==23.5.26
folium==0.14.0
fonttools==4.25.0
fpdf==1.7.2
frozenlist==1.4.0
fsspec==2023.10.0
funcy==2.0
gast==0.4.0
gensim==4.3.2
gitdb==4.0.10
GitPython==3.1.30
google==3.0.0
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-play-scraper==1.2.6
googleapis-common-protos==1.63.1
googletrans==3.1.0a0
graphviz==0.20.1
grpcio==1.56.2
gspread==5.12.4
gspread-dataframe==4.0.0
gspread-formatting==1.2.0
gspread-pandas==3.3.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
h5py==3.11.0
happytransformer==2.4.1
hmmlearn==0.3.2
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.23.4
humanize==4.7.0
hyperframe==5.2.0
idna==2.10
imbalanced-learn==0.11.0
inflection==0.5.1
Jinja2==3.0.1
kaleido==0.2.1
keras==3.4.1
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.2.post1
Mako==1.3.0
markdown-it-py==2.1.0
mccabe==0.6.1
mdurl==0.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
ml-dtypes==0.3.2
mlxtend==0.23.0
mpmath==1.2.1
msgpack==1.0.8
multidict==6.0.4
multiprocess==0.70.15
munkres==1.1.4
murmurhash==1.0.9
mypy-extensions==0.4.3
mysql-connector-python==8.2.0
mysqlclient==2.2.0
namex==0.0.8
navigator-updater==0.3.0
neattext==0.1.3
numexpr==2.10.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.0.76
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
outcome==1.2.0
packaging==23.2
pandas==1.5.3
pathy==0.10.2
patsy==0.5.2
pdfkit==1.0.0
pep8==1.7.1
Pillow==9.4.0
platformdirs==4.2.2
plotly-express==0.4.1
polib==1.2.0
pooch==1.8.1
preshed==3.0.8
proto-plus==1.23.0
protobuf==4.21.12
psycopg2==2.9.9
pyarrow==11.0.0
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.22
pycurl==7.45.1
pydantic==1.10.12
pydeck==0.8.0
PyDispatcher==2.0.5
pydub==0.25.1
pyfiglet==1.0.2
Pygments==2.14.0
pygwalker==0.3.2
pyLDAvis==3.4.1
pyls-spyder==0.4.0
Pympler==1.0.1
PyMySQL==1.1.1
pyswarm==0.6
pytest==7.1.3
python-dotenv==0.21.0
python-lsp-jsonrpc==1.0.0
pytz-deprecation-shim==0.1.0.post0
pywin32==302
PyYAML==6.0
queuelib==1.5.0
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rfc3986==1.5.0
rich==13.3.1
rsa==4.9
safetensors==0.4.3
Sastrawi==1.0.1
scikit-learn==1.3.1
scikit-learn-intelex==2021.20221004.171935
scipy==1.9.1
selenium==4.8.3
semver==2.13.0
sentencepiece==0.1.99
sip==4.19.13
smmap==5.0.0
snscrape==0.7.0.20230622
soundfile==0.12.1
soxr==0.3.7
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sql_metadata==2.12.0
SQLAlchemy==1.4.0
sqlparse==0.5.1
srsly==2.4.7
st-gsheets-connection==0.0.4
statsmodels==0.13.2
streamlit==1.38.0
streamlit-option-menu==0.3.2
swifter==1.4.0
tables==3.6.1
TBB==0.2
tensorboard==2.16.2
tensorboard-data-server==0.7.1
tensorflow==2.16.2
tensorflow-estimator==2.13.0
tensorflow-intel==2.16.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
textblob==0.17.1
thinc==8.1.10
threadpoolctl==3.5.0
tkcalendar==1.6.1
tokenizers==0.19.1
torch==2.1.1
torchaudio==2.1.1
torchsummary==1.5.1
torchvision==0.16.1
transformers==4.41.2
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.14.0
typer==0.9.0
typing_extensions==4.12.2
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.19
vaderSentiment==3.3.2
validators==0.22.0
virtualenv==20.17.1
wasabi==1.1.2
webencodings==0.5.1
Werkzeug==2.0.1
win-unicode-console==0.5
wincertstore==0.2
wordcloud==1.9.2
wsproto==1.2.0
xgboost==1.7.6
xxhash==3.3.0
yarl==1.9.2
zict==2.1.0
